moved illus
