+++ frontend/src/atoms/illustrations/Error404.tsx
