+++ frontend/src/atoms/illustrations/cardBackground.tsx
