+++ frontend/src/components/Cards.tsx
@@ -1,5 +1,5 @@
 import { ReactNode } from "react";
-import { CardBackground } from "./illustrations/cardBackground";
+import { CardBackground } from "../atoms/illustrations/cardBackground";
 import { useRoleTheme } from "../context/roleThemeContext";
 
 export type CardVariant = "projects" | "deliverables" | "tasks" | "toReview";
